backend/requirements.txt
Django==3.2.8
django-allauth==0.54.0
django-cors-headers==4.3.1
django-environ==0.9.0
djangorestframework==3.14.0
Flask==2.3.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
django-phonenumber-field[phonenumbers]